+++ adapters/Pangle/PangleAdapter/GADMAdapterPangleUtils.m
@@ -38,6 +38,97 @@ void GADMAdapterPangleMutableSetAddObject(NSMutableSet *_Nullable set, NSObject 
   }
 }
 
+GADMAdapterPangleConsentResult GADMAdapterPangleHasACConsent(NSInteger vendorId) {
+  NSUserDefaults *userDefaults = [NSUserDefaults standardUserDefaults];
+
+  NSInteger gdprApplies = [userDefaults integerForKey:@"IABTCF_gdprApplies"];
+  if (gdprApplies != 1) {
+    return GADMAdapterPangleConsentResultUnknown;
+  }
+
+  NSString *additionalConsentString = [userDefaults stringForKey:@"IABTCF_AddtlConsent"];
+  if (!additionalConsentString.length) {
+    return GADMAdapterPangleConsentResultUnknown;
+  }
+
+  NSString *vendorIdString = @(vendorId).stringValue;
+  NSArray<NSString *> *additionalConsentParts =
+      [additionalConsentString componentsSeparatedByString:@"~"];
+
+  NSInteger version = additionalConsentParts[0].integerValue;
+  if (version == 1) {
+    // Spec version 1
+    NSLog(@"The IABTCF_AddtlConsent string uses version 1 of Google’s Additional Consent "
+           "spec. Version 1 does not report vendors to whom the user denied consent. To "
+           "detect vendors that the user denied consent, upgrade to a CMP that supports "
+           "version 2 of Google's Additional Consent technical specification.");
+
+    if (additionalConsentParts.count == 1) {
+      // The AC string had no consented vendor.
+      return GADMAdapterPangleConsentResultUnknown;
+    } else if (additionalConsentParts.count == 2) {
+      NSArray<NSString *> *consentedIds =
+          [additionalConsentParts[1] componentsSeparatedByString:@"."];
+      if ([consentedIds containsObject:vendorIdString]) {
+        return GADMAdapterPangleConsentResultTrue;
+      }
+
+      return GADMAdapterPangleConsentResultUnknown;
+    } else {
+      NSString *errorMessage =
+          [NSString stringWithFormat:
+                        @"Could not parse the IABTCF_AddtlConsent string: \"%@\". String had more "
+                        @"parts than expected. Did your CMP write IABTCF_AddtlConsent correctly?",
+                        additionalConsentString];
+      NSLog(@"%@", errorMessage);
+      return GADMAdapterPangleConsentResultUnknown;
+    }
+  } else if (version >= 2) {
+    // Spec version 2 and above.
+    if (additionalConsentParts.count < 3) {
+      NSString *errorMessage =
+          [NSString stringWithFormat:
+                        @"Could not parse the IABTCF_AddtlConsent string: \"%@\". String has less "
+                        @"parts than expected. Did your CMP write IABTCF_AddtlConsent correctly?",
+                        additionalConsentString];
+      NSLog(@"%@", errorMessage);
+      return GADMAdapterPangleConsentResultUnknown;
+    }
+
+    NSArray<NSString *> *disclosedIds =
+        [additionalConsentParts[2] componentsSeparatedByString:@"."];
+    if (![disclosedIds[0] isEqualToString:@"dv"]) {
+      NSString *errorMessage = [NSString
+          stringWithFormat:
+              @"Could not parse the IABTCF_AddtlConsent string: \"%@\". Expected disclosed vendors "
+              @"part to have the string \"dv.\". Did your CMP write IABTCF_AddtlConsent correctly?",
+              additionalConsentString];
+      NSLog(@"%@", errorMessage);
+      return GADMAdapterPangleConsentResultUnknown;
+    }
+
+    NSArray<NSString *> *consentedIds =
+        [additionalConsentParts[1] componentsSeparatedByString:@"."];
+    if ([consentedIds containsObject:vendorIdString]) {
+      return GADMAdapterPangleConsentResultTrue;
+    }
+
+    if ([disclosedIds containsObject:vendorIdString]) {
+      return GADMAdapterPangleConsentResultFalse;
+    }
+
+    return GADMAdapterPangleConsentResultUnknown;
+  } else {
+    // Unknown spec version
+    NSString *errorMessage = [NSString
+        stringWithFormat:@"Could not parse the IABTCF_AddtlConsent string: \"%@\". Spec version "
+                         @"was unexpected. Did your CMP write IABTCF_AddtlConsent correctly?",
+                         additionalConsentString];
+    NSLog(@"%@", errorMessage);
+    return GADMAdapterPangleConsentResultUnknown;
+  }
+}
+
 @implementation GADMAdapterPangleUtils
 
 + (BOOL)isChildUser {